+++ lib/volume_controller.dart
@@ -1,1 +1,4 @@
+/// A Flutter plugin to control system volume on Android, iOS, Windows, macOS, and Linux.
+library;
+
 export 'src/volume_controller.dart';